Factor clipboard logic out of Editor.ts
